% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STA523_plot_density.R
\name{g_legend}
\alias{g_legend}
\title{A function that extract theme}
\usage{
g_legend(a.gplot)
}
\arguments{
\item{a.gplot}{A ggplot object}
}
\value{
return a legend
}
\description{
An internal function that extract theme, only for plot arrangement purpose
}
