% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STA523_Model.R
\name{logit}
\alias{logit}
\title{Logit Function}
\usage{
logit(x)
}
\arguments{
\item{x}{A numeric vector input to put through logistic function}
}
\value{
(log(x/(1-x)))
}
\description{
Logit Function for internal predictions
}
