% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STA523_Model.R
\name{MSE}
\alias{MSE}
\title{MSE Function}
\usage{
MSE(y, y_hat)
}
\arguments{
\item{y}{A numeric input (of vector) of true y}

\item{y_hat}{A numeric input (of vector) of predicted y}
}
\value{
MSE functio nreturn a numeric value indicating the mean square error of the input.
}
\description{
This function calculate the Mean Squared Error of of the the predicted y value and the true y value.
}
