% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STA523_plot_cor.R
\name{plot_cor}
\alias{plot_cor}
\title{plot the gradApplication dataset pairwise scatter plot}
\usage{
plot_cor()
}
\value{
return Nothing
}
\description{
plot the correlation of gradApplication dataset for EDA purpose
}
