% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STA523_plot_density.R
\name{plot_density}
\alias{plot_density}
\title{plot the gradApplication dataset density distribution}
\usage{
plot_density()
}
\value{
return Nothing
}
\description{
plot the gradApplication dataset for EDA purpose
}
