% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STA523_Model.R
\name{predict_skr}
\alias{predict_skr}
\title{Predict Function}
\usage{
predict_skr(model, newx, method)
}
\arguments{
\item{model}{An inputted model}

\item{newx}{inputted data to predict}

\item{method}{The type of the model}
}
\description{
A general wrapper function to predict using the inputted model
}
