% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI&Server.R
\name{Server}
\alias{Server}
\title{Shiny Server}
\usage{
Server(input, output, session)
}
\arguments{
\item{input}{shiny inputs}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
This function provides server part of our Shiny App
}
