% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STA523_Model.R
\name{split_data}
\alias{split_data}
\title{Split Train Test Function}
\usage{
split_data(data = gradApplication, train_p = 0.8, seed = 0)
}
\arguments{
\item{data}{splitting data, default to Graduate Admission data set}

\item{train_p}{Proportion of the training set (out of the entire dataset) default 0.8}

\item{seed}{seeding during random sampling, default 0}
}
\value{
A list containing two dataframe called train and test.
}
\description{
Intake a data file path, split it into train test set by user input proportion
}
