% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STA523_Model.R
\name{train_model}
\alias{train_model}
\title{Train Model Function}
\usage{
train_model(traindata, method = "ols")
}
\arguments{
\item{traindata}{Intake a dataframe as training dataset}

\item{method}{Intake a string to indicate the method of modeling}
}
\value{
A model trained using the train data
}
\description{
This function is a wrapper function to training the model that
predict the acceptance rate of the prospective student.Currently we are supporting OLS,
Ridge and Lasso. CV tuning for Ridge and Lasso are already implemented in this function
}
